Add solution to css
implements solution for 02 css exercise.
some improvements could be done to clarity by repeating the margin
property in each class instead of trying to set a global setting and
for this application in should be ok to do so. i dediced to leave it
like this for better perspective on the progress. aside everything is
the same, the provided solution prefers a margin-bottom inseatead but
it doesn't really matter to meeeee