Add solution 02-flex
nav bar using flex box, all quite easy aside for the padding
initially applied to the ul instead of the .header, where it makes
more sense to be, aside put the padding and margin to 0 for the ul
elements. checked with provided solution, more care in regards with
padding and where it makes more sense to be put.