Add solution 07 flex DONE :)
all good, got the hang of how it works at this level, some mistakes
in thinking which elements interact with others in what way but
nothing major. some stuff is different from the provided solution
but its mostly ease of use and some oddities. should check how
to position correctly and in which order the various style descr.
in the .css file and how to properly name classes in html